review fixes: general improvements
