+++ server/server.cmake
@@ -24,7 +24,6 @@ prepend(KPHP_TASK_WORKERS_SOURCES ${BASE_DIR}/server/task-workers/
         task-workers-context.cpp
         pipe-io.cpp
         pending-tasks.cpp
-        task-workers.cpp
         shared-context.cpp
         shared-memory-manager.cpp)
